+++ sdk/appconfiguration/app-configuration/src/internal/helpers.ts
@@ -14,10 +14,12 @@ import {
   ConfigurationSnapshot,
   SnapshotResponse,
   EtagEntity,
+  ListLabelsOptions,
 } from "../models";
 import { FeatureFlagHelper, FeatureFlagValue, featureFlagContentType } from "../featureFlag";
 import {
   GetKeyValuesOptionalParams,
+  GetLabelsOptionalParams,
   GetSnapshotsOptionalParams,
   KeyValue,
 } from "../generated/src/models";
@@ -44,6 +46,13 @@ export interface SendConfigurationSettingsOptions
    */
   snapshotName?: string;
 }
+
+/**
+ * Options for listLabels that allow for filtering based on keys, labels and other fields.
+ * Also provides `fields` which allows you to selectively choose which fields are populated in the
+ * result.
+ */
+export interface SendLabelsRequestOptions extends ListLabelsOptions {}
 
 /**
  * Formats the etag so it can be used with a If-Match/If-None-Match header
@@ -113,7 +122,7 @@ export function checkAndFormatIfAndIfNoneMatch(
  */
 export function formatFiltersAndSelect(
   listConfigOptions: ListRevisionsOptions,
-): Pick<GetKeyValuesOptionalParams, "key" | "label" | "select" | "acceptDatetime"> {
+): Pick<GetKeyValuesOptionalParams, "key" | "label" | "select" | "acceptDatetime" | "tags"> {
   let acceptDatetime: string | undefined = undefined;
 
   if (listConfigOptions.acceptDateTime) {
@@ -122,6 +131,7 @@ export function formatFiltersAndSelect(
   return {
     key: listConfigOptions.keyFilter,
     label: listConfigOptions.labelFilter,
+    tags: listConfigOptions.tagsFilter,
     acceptDatetime,
     select: formatFieldsForSelect(listConfigOptions.fields),
   };
@@ -138,7 +148,10 @@ export function formatFiltersAndSelect(
  */
 export function formatConfigurationSettingsFiltersAndSelect(
   listConfigOptions: SendConfigurationSettingsOptions,
-): Pick<GetKeyValuesOptionalParams, "key" | "label" | "select" | "acceptDatetime" | "snapshot"> {
+): Pick<
+  GetKeyValuesOptionalParams,
+  "key" | "label" | "select" | "acceptDatetime" | "snapshot" | "tags"
+> {
   const { snapshotName: snapshot, ...options } = listConfigOptions;
   return {
     ...formatFiltersAndSelect(options),
@@ -160,6 +173,23 @@ export function formatSnapshotFiltersAndSelect(
     name: listSnapshotOptions.nameFilter,
     status: listSnapshotOptions.statusFilter,
     select: listSnapshotOptions.fields,
+  };
+}
+
+/**
+ * Transforms some of the key fields in ListLabelsOptions
+ * so they can be added to a request using AppConfigurationGetLabelsOptionalParams.
+ * - `select` is populated with the proper field names from `options.fields`
+ * - `nameFilter` are moved to name
+ *
+ * @internal
+ */
+export function formatLabelsFiltersAndSelect(
+  listLabelsOptions: ListLabelsOptions,
+): Pick<GetLabelsOptionalParams, "name" | "select"> {
+  return {
+    name: listLabelsOptions.nameFilter,
+    select: listLabelsOptions.fields,
   };
 }
 /**