+++ sdk/batch/batch-rest/test/utils/env.ts
@@ -1,0 +1,6 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+import * as dotenv from "dotenv";
+
+dotenv.config();