+++ sdk/maps/maps-render-rest/test/public/mapsRender.spec.ts
@@ -8,7 +8,6 @@ import { assert } from "chai";
 import { createClient, createRecorder } from "./utils/recordedClient";
 import { Context } from "mocha";
 import MapsRender, { isUnexpected, MapsRenderClient } from "../../src";
-import { AzureKeyCredential } from "@azure/core-auth";
 
 describe("Authentication", function () {
   let recorder: Recorder;
@@ -21,15 +20,7 @@ describe("Authentication", function () {
     await recorder.stop();
   });
 
-  it("should work with Shared Key authentication", async function () {
-    const credential = new AzureKeyCredential(env["MAPS_SUBSCRIPTION_KEY"] as string);
-    const client = MapsRender(credential, recorder.configureClientOptions({}));
-
-    const response = await client.path("/map/copyright/caption/{format}", "json").get();
-    assert.isOk(!isUnexpected(response));
-  });
-
-  it("should work with AAD authentication", async function () {
+  it("should work with Microsoft Entra ID authentication", async function () {
     /**
      * Skip this test in browser because we have to use InteractiveBrowserCredential in the browser.
      * But it requires user's interaction, which is not testable in karma.
@@ -197,7 +188,6 @@ describe("MapsRender", () => {
       assert.fail(response.body.error?.message || "Unexpected error.");
     }
 
-    assert.isNotEmpty(response.body.tilejson);
     assert.isNotEmpty(response.body.tiles);
   });
 });