+++ sdk/monitor/monitor-opentelemetry-exporter/src/sampling.ts
@@ -20,7 +20,7 @@ export class ApplicationInsightsSampler implements Sampler {
    */
   constructor(samplingRatio: number = 1) {
     this.samplingRatio = samplingRatio;
-    if (this.samplingRatio > 1) {
+    if (this.samplingRatio > 1 || this.samplingRatio < 0 || !Number.isFinite(this.samplingRatio)) {
       throw new Error("Wrong sampling rate, data will not be sampled out");
     }
     this._sampleRate = Math.round(this.samplingRatio * 100);
@@ -41,15 +41,15 @@ export class ApplicationInsightsSampler implements Sampler {
    * @returns a {@link SamplingResult}.
    */
   public shouldSample(
-    // @ts-ignore
+    // @ts-expect-error unused argument
     context: Context,
     traceId: string,
-    // @ts-ignore
+    // @ts-expect-error unused argument
     spanName: string,
-    // @ts-ignore
+    // @ts-expect-error unused argument
     spanKind: SpanKind,
     attributes: Attributes,
-    // @ts-ignore
+    // @ts-expect-error unused argument
     links: Link[],
   ): SamplingResult {
     let isSampledIn = false;
@@ -62,7 +62,11 @@ export class ApplicationInsightsSampler implements Sampler {
     }
     // Add sample rate as span attribute
     attributes = attributes || {};
-    attributes[AzureMonitorSampleRate] = this._sampleRate;
+
+    // Only send the sample rate if it's not 100
+    if (this._sampleRate !== 100) {
+      attributes[AzureMonitorSampleRate] = this._sampleRate;
+    }
     return isSampledIn
       ? { decision: SamplingDecision.RECORD_AND_SAMPLED, attributes: attributes }
       : { decision: SamplingDecision.NOT_RECORD, attributes: attributes };