+++ sdk/storage/storage-queue/src/QueueSASSignatureValues.ts
@@ -80,6 +80,14 @@ export function generateQueueSASQueryParameters(
   queueSASSignatureValues: QueueSASSignatureValues,
   sharedKeyCredential: StorageSharedKeyCredential,
 ): SASQueryParameters {
+  return generateQueueSASQueryParametersInternal(queueSASSignatureValues, sharedKeyCredential)
+    .sasQueryParameters;
+}
+
+export function generateQueueSASQueryParametersInternal(
+  queueSASSignatureValues: QueueSASSignatureValues,
+  sharedKeyCredential: StorageSharedKeyCredential,
+): { sasQueryParameters: SASQueryParameters; stringToSign: string } {
   if (
     !queueSASSignatureValues.identifier &&
     !(queueSASSignatureValues.permissions && queueSASSignatureValues.expiresOn)
@@ -119,18 +127,21 @@ export function generateQueueSASQueryParameters(
 
   const signature = sharedKeyCredential.computeHMACSHA256(stringToSign);
 
-  return new SASQueryParameters(
-    version,
-    signature,
-    verifiedPermissions,
-    undefined,
-    undefined,
-    queueSASSignatureValues.protocol,
-    queueSASSignatureValues.startsOn,
-    queueSASSignatureValues.expiresOn,
-    queueSASSignatureValues.ipRange,
-    queueSASSignatureValues.identifier,
-  );
+  return {
+    sasQueryParameters: new SASQueryParameters(
+      version,
+      signature,
+      verifiedPermissions,
+      undefined,
+      undefined,
+      queueSASSignatureValues.protocol,
+      queueSASSignatureValues.startsOn,
+      queueSASSignatureValues.expiresOn,
+      queueSASSignatureValues.ipRange,
+      queueSASSignatureValues.identifier,
+    ),
+    stringToSign: stringToSign,
+  };
 }
 
 function getCanonicalName(accountName: string, queueName: string): string {