+++ sdk/monitor/monitor-opentelemetry/CHANGELOG.md
@@ -14,6 +14,8 @@
 ### Other Changes
 
 - Update applicationinsights-web-snippet version to 1.2.1.
+- Add statsbeat tracking for all OpenTelemetry instrumentations.
+- Update value used to send sample rate to ingestion.
 
 ## 1.6.0 (2024-06-13)